web/just_l33t/src/backend/requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Jinja2==3.1.2
mysql-connector-python==8.0.32
PyJWT==2.6.0
uWSGI==2.0.21
Werkzeug==2.2.3